+++ packages/f2/package.json
@@ -4,6 +4,7 @@
   "main": "lib/index.js",
   "module": "es/index.js",
   "types": "es/index.d.ts",
+  "miniprogram": "dist",
   "keywords": [
     "f2",
     "charts",
