+++ packages/wx/examples/test/package.json
@@ -3,9 +3,8 @@
     "beforeCompile": "babel pages --out-dir pages --only **/*.jsx"
   },
   "dependencies": {
-    "@antv/f2": "~4.0.0-alpha.3",
-    "@antv/f2-wx": "~4.0.0-alpha.0",
-    "@babel/runtime": "^7.12.5"
+    "@antv/f2": "^4.0.0-alpha.7",
+    "@antv/f2-wx": "~4.0.0-alpha.0"
   },
   "devDependencies": {
     "@babel/cli": "^7.16.0",
